update with nymph

+++ tools/rename_copy_files.py
@@ -2,6 +2,10 @@ import os
 import json
 from pathlib import Path
 
+def files_binary_equal(full_path, full_new_path):
+    with open(full_path, 'rb') as f1, open(full_new_path, 'rb') as f2:
+        return f1.read() == f2.read()
+
 def rename_copy_files(directory):
     num_renamed = 0
     num_of_files = 0
@@ -12,11 +16,20 @@ def rename_copy_files(directory):
 
             for num in range(5):
                 num_suffix = f" ({num})"
-                if num_suffix in file:
-                    num_renamed += 1
+                if num_suffix in file:                    
                     new_file = file.replace(num_suffix, "")
-                    full_new_path = Path(os.path.join(root, new_file)).absolute()                   
+                    full_new_path = Path(os.path.join(root, new_file)).absolute()
+                    
+                    if full_new_path.exists():
+                        if not files_binary_equal(full_path, full_new_path):
+                            # If the destination path already exists, check if
+                            # the file is just a duplicate. If it is then just skip
+                            # else we should throw an error
+                            print(f"\nFile already exists: {full_new_path}")
+                        continue
+
                     print(full_path, full_new_path)
+                    num_renamed += 1
                     os.rename(full_path, full_new_path)
 
     print("Renamed {}/{} files".format(num_renamed, num_of_files))
